% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baxter.R
\name{get_metadata}
\alias{get_metadata}
\title{Reads and cleans raw metadata}
\usage{
get_metadata()
}
\value{
The cleaned metadata as a dataframe
}
\description{
Reads and cleans raw metadata
}
\examples{
metadata <- get_metadata()
}
